\name{make_qPCR_plots}
\alias{make_qPCR_plots}
\title{Creates Fig 4, the qPCR plots}
\usage{
  make_qPCR_plots(all_data, corrected_seds, ps = 0.75)
}
\arguments{
  \item{all_data}{The `all_data` data frame}

  \item{corrected_seds}{The `corrected_seds` data frame.
  Yes, these are redundant}

  \item{ps}{The point size for the plots b and c}
}
\description{
  Creates Fig 4, the qPCR plots
}

