\name{make_sed_yield_boxplots}
\alias{make_sed_yield_boxplots}
\title{Make boxplots of yield for sediments data, split by various factors}
\usage{
  make_sed_yield_boxplots(corrected_seds, n = 1000)
}
\arguments{
  \item{corrected_seds}{Data frame of corrected sediments}

  \item{n}{Number of replicates to be passed on to
  `aov_perm_test()`}
}
\description{
  Make boxplots of yield for sediments data, split by
  various factors
}

