\name{plot_cell_vs_fish}
\alias{plot_cell_vs_fish}
\title{Make a plot of totalcells vs CARDFISH (or other FISH) counts in seawater & sediments}
\usage{
  plot_cell_vs_fish(allData, point_size = 0.5)
}
\arguments{
  \item{allData}{The data frame of all data from
  merge_seds_and_sw.R}

  \item{point_size}{The size of points on the printed plot}
}
\value{
  Returns the ggplot object
}
\description{
  Make a plot of totalcells vs CARDFISH (or other FISH)
  counts in seawater & sediments
}

