\name{sed_bac_and_arc_v_depth}
\alias{sed_bac_and_arc_v_depth}
\title{Makes plots and models of sediments bac and arc concentrations vs depth}
\usage{
  sed_bac_and_arc_v_depth(corrected_sed, point_size = 0.75,
    reproduce_exactly = FALSE)
}
\arguments{
  \item{corrected_sed}{Data frame of sediments data}

  \item{point_size}{Point size for the plot}

  \item{reproduce_exactly}{If TRUE, set the state of the
  random number generator (set.seed(2112))}
}
\description{
  Makes plots and models of sediments bac and arc
  concentrations vs depth
}

