\name{sed_depth_trends}
\alias{sed_depth_trends}
\title{plots depth trends}
\usage{
  sed_depth_trends(corrected_seds, point_size = 0.5,
    title_font_size = 7)
}
\arguments{
  \item{corrected_seds}{The corrected seds data frame}

  \item{point_size}{For both the scatterplots and the
  outliers in the box plots}

  \item{title_font_size}{For all three plots}
}
\description{
  plots depth trends
}
\details{
  Note that this function is currently very messy, and
  needs to be cleaned up
}

