\name{significance_labeller}
\alias{significance_labeller}
\title{Create 'significance code' based on p values}
\usage{
  significance_labeller(p, alpha = 0.05, sig_text = "S",
    nonsig_text = "NS", print_p = TRUE, sig_figs = 2,
    tiny = 0.001)
}
\description{
  Create 'significance code' based on p values
}

