\name{single_sw_yield_boxplot}
\alias{single_sw_yield_boxplot}
\title{Make a single yield boxplot (seds and sw, despite the name)}
\usage{
  single_sw_yield_boxplot(d, xvar, yvar, sig_text = NA,
    jitter_width = 0.25, alpha = 0.25, ps = 0.75,
    text_size = 4)
}
\description{
  Function to make generic boxplots as in the supplemental
  figures
}

