\name{sw_depth_profiles}
\alias{sw_depth_profiles}
\title{Plots depth profiles of various things in seawater}
\usage{
  sw_depth_profiles(corrected_sw, point_size = 0.75)
}
\arguments{
  \item{corrected_sw}{Data frame of seawater data}

  \item{point_size}{point size in plot}
}
\description{
  Creates depth profiles in figures 5 a-c: Bacteria,
  Archaea, and %Archaea as a function of depth in seawater.
  Also performs breakpoint analysis
}

